+++ dflib-echarts/src/test/java/org/dflib/echarts/EChartHtmlSaverTest.java
@@ -30,8 +30,27 @@ public class EChartHtmlSaverTest {
         StringWriter out = new StringWriter();
         ECharts.saver().title("Test Charts").save(out, p1, p2);
 
-
         String expected = new String(getClass().getResourceAsStream("charts1.html").readAllBytes(), StandardCharsets.UTF_8);
+        assertEquals(expected, out.toString());
+    }
+
+    @Test
+    public void embeddedTemplate() throws IOException {
+        ElementIdGenerator idGenerator = ElementIdGenerator.sequential();
+
+        EChartHtml p1 = ECharts
+                .chart(idGenerator)
+                .xAxis("x")
+                .series(SeriesOpts.ofLine().name("Y1"), "y1").plot(df2);
+        EChartHtml p2 = ECharts
+                .chart(idGenerator)
+                .xAxis("x")
+                .series(SeriesOpts.ofLine().name("Y2"), "y2").plot(df2);
+
+        StringWriter out = new StringWriter();
+        ECharts.saver().defaultEmbeddedTemplate().title("Test Charts").save(out, p1, p2);
+
+        String expected = new String(getClass().getResourceAsStream("charts-embedded1.html").readAllBytes(), StandardCharsets.UTF_8);
         assertEquals(expected, out.toString());
     }